+++ content/sections/2-Projects.ja.md
@@ -1,0 +1,55 @@
+---
+anchor: "ポートフォリオ"
+header: "ポートフォリオ"
+subheader: "あなたは偶然断然その持通り"
+portfolios: [
+  {
+    imageFileName: "portfolio/01-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/01-full.jpg",
+    header: "Threads",
+    subheader: "イラストレーション",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Threads", "カテゴリ: Illustration"]
+  },
+  {
+    imageFileName: "portfolio/02-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/02-full.jpg",
+    header: "Explorer",
+    subheader: "グラフデザイン",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Explorer", "カテゴリ: Graphic Design"]
+  },
+  {
+    imageFileName: "portfolio/03-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/03-full.jpg",
+    header: "Finish",
+    subheader: "アイデンティティ",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Finish", "カテゴリ: Identity"]
+  },
+  {
+    imageFileName: "portfolio/04-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/04-full.jpg",
+    header: "Lines",
+    subheader: "ブランディング",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Lines", "カテゴリ: Branding"]
+  },
+  {
+    imageFileName: "portfolio/05-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/05-full.jpg",
+    header: "Southwest",
+    subheader: "ウェブサイトデザイン",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Southwest", "カテゴリ: Website Design"]
+  },
+  {
+    imageFileName: "portfolio/06-thumbnail.jpg",
+    imageFileNameDetail: "portfolio/06-full.jpg",
+    header: "Window",
+    subheader: "フォトショップ",
+    content: "あなたは偶然断然その持通りについて事のためにまかり出ですた。もう元来を説明心はとうとうこのお話しないななどでいて行くたでは滅亡しましでて、再びにはなったうないです。受売へ折っだっのはもし今をもうないないない。",
+    extraInfo: ["日付： January 2017", "クライエント: Window", "カテゴリ: Photography"]
+  }
+]
+---
