--- src/components/PortfolioItem.scss
@@ -1,63 +1,0 @@
-.portfolio-item {
-  right: 0;
-  margin: 0 0 15px;
-  .portfolio-link {
-    position: relative;
-    display: block;
-    max-width: 400px;
-    margin: 0 auto;
-    cursor: pointer;
-    .portfolio-hover {
-      position: absolute;
-      top: 0;
-      width: 100%;
-      height: 100%;
-      transition: all ease 0.5s;
-      opacity: 0;
-      background: fade-out($primary, 0.1);
-      &:hover {
-        opacity: 1;
-      }
-      .portfolio-hover-content {
-        font-size: 20px;
-        position: absolute;
-        top: 50%;
-        width: 100%;
-        height: 20px;
-        margin-top: -12px;
-        text-align: center;
-        color: white;
-        i {
-          margin-top: -12px;
-        }
-        h3,
-        h4 {
-          margin: 0;
-        }
-      }
-    }
-  }
-  .portfolio-caption {
-    max-width: 400px;
-    margin: 0 auto;
-    padding: 25px;
-    text-align: center;
-    background-color: $white;
-    h4 {
-      margin: 0;
-      text-transform: none;
-    }
-    p {
-      font-size: 16px;
-      font-style: italic;
-      margin: 0;
-      @include serif-font;
-    }
-  }
-}
-
-@media (min-width: 767px) {
-  .portfolio-item {
-    margin: 0 0 30px;
-  }
-}
