--- src/views/Sections/Portfolio/Portfolio.scss
@@ -1,6 +1,0 @@
-// Styling for the portfolio section
-.portfolio-section {
-  * {
-    z-index: 2;
-  }
-}
