--- src/views/Sections/Portfolio/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Portfolio";
