+++ src/views/Sections/Projects/Projects.scss
@@ -1,0 +1,6 @@
+// Styling for the portfolio section
+.portfolio-section {
+  * {
+    z-index: 2;
+  }
+}
