+++ src/views/Sections/index.js
@@ -1,6 +1,6 @@
 export { default as Experience } from "./Experience";
 export { default as Contact } from "./Contact";
-export { default as Portfolio } from "./Portfolio";
+export { default as Projects } from "./Projects";
 export { default as Services } from "./Services";
 export { default as Me } from "./Me";
 export { default as Publications } from "./Publications";
